Validation

+++ web/andy/andy_to_shelf.php
@@ -1,4 +1,5 @@
 <?php
+require 'db_conn.php';
 $username = $_POST['user_id'];
 $isbn = $_POST['isbn'];
 $add = $_POST['add'];
@@ -9,11 +10,17 @@ if($add=='true')
     $query = "insert into book_shelf values ('$username','$isbn');";
     $result = $pg_query($query);
 }
-else
+if($add=='false')
 {
     $query = "delete from book_shelf where user_id='$username' and isbn='$isbn';";
     $result = $pg_query($query);
 }
-
-echo json_encode(array('done_query'=>'true'));
+if($result)
+{
+  echo json_encode(array('done_query'=>'true'));
+}
+else
+{
+  echo json_encode(array('done_query'=>'false'));
+}
 ?>
